added unit tests for the simple serialization only *passing capabilities
Also added a helper file testConnection.js which hooks up the loopback and gives you the connection promise
