+++ test/coterminous-spec.js
@@ -6,7 +6,6 @@ import {
 from '../src/coterminous.js';
 
 import Coterminous from '../publish.js';
-import Loopback from '../bower_components/coterminous-js-loopback/bower/loopback.js';
 
 var mockedCapability = {
   "name": "mocked",
@@ -39,31 +38,4 @@ describe('Coterminous', () => {
     });
 
   });
-
-  describe('connect test', () => {
-    it('ensure able to define root', () => {
-      Coterminous.root({
-        hello: function() {
-          return "world";
-        }
-      });
-    });
-    it('ensure root definition available from other side', (done) => {
-      Coterminous.root({
-        hello: function() {
-          return "world";
-        }
-      });
-      
-      var transport = new Loopback();
-
-      Coterminous.connect(transport.A).then(function(remoteInterface) {
-        assert.equal(typeof remoteInterface.hello, 'function', 'Root remote interface was not passed correctly');
-        done();
-      });
-      Coterminous.connectTransport(transport.B);
-    });
-  });
-
-
 });
